refactor key.js
